can test boss damage

+++ TestScene/Assets/_Dev 1/AiProgramming_Department/AI/Scripts/AICharacter.cs
@@ -196,7 +196,7 @@ public class AICharacter : MonoBehaviour
         return player.transform.position;
     }
 
-    public int GetHealth()
+    public float GetHealth()
     {
         return this.health;
     }
